LLM/email-agent/pyproject.toml
[project]
name = "local-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imap-tools>=1.10.0",
    "langchain>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.1.5",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
